+++ CS-MySQL-Library/MySQL/MySQLResult.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections.Generic;
+
+namespace GHMatti.MySQL
+{
+    // Renaming the Result from Queries to something shorter
+    public class MySQLResult : List<Dictionary<string, dynamic>> { }
+}